Big refactoring
Refactored LockedTranslationUnit and Cache into TranslationUnit.
Locking is now private to the class. Added Diagnostics class.

Moved compile argument collection out of TranslationCache and
into the plugin only. Also started using CompileOptions class.

Removed a bunch of useless settings.

Added double underscores to private class method names.
